refactor: factor out optimisation structures
