+++ pumpkin-solver/src/api/outputs/mod.rs
@@ -47,7 +47,7 @@ pub enum SatisfactionResultUnderAssumptions<'solver, 'brancher, B: Brancher> {
     Unknown,
 }
 
-/// The result of a call to [`Solver::maximise`] or [`Solver::minimise`].
+/// The result of a call to [`Solver::optimise`].
 #[derive(Debug)]
 pub enum OptimisationResult {
     /// Indicates that an optimal solution has been found and proven to be optimal. It provides an
