+++ pumpkin-solver/src/lib.rs
@@ -103,8 +103,9 @@
 //! }
 //! ```
 //!
-//! **Optimizing an objective** can be done in a similar way using [`Solver::maximise`] or
-//! [`Solver::minimise`]; first the objective variable and a constraint over this value are added:
+//! **Optimizing an objective** can be done in a similar way using [`Solver::optimise`]; first the
+//! objective variable and a constraint over this value are added:
+//!
 //! ```rust
 //! # use pumpkin_solver::Solver;
 //! # use pumpkin_solver::constraints;
@@ -122,7 +123,7 @@
 //!     .post();
 //! ```
 //!
-//! Then we can find the optimal solution using [`Solver::minimise`] or [`Solver::maximise`]:
+//! Then we can find the optimal solution using [`Solver::optimise`]:
 //! ```rust
 //! # use pumpkin_solver::Solver;
 //! # use pumpkin_solver::results::OptimisationResult;
@@ -130,6 +131,8 @@
 //! # use pumpkin_solver::results::ProblemSolution;
 //! # use pumpkin_solver::constraints;
 //! # use pumpkin_solver::constraints::Constraint;
+//! # use pumpkin_solver::optimisation::OptimisationDirection;
+//! # use pumpkin_solver::optimisation::UpperBoundingSearch;
 //! # use std::cmp::max;
 //! # let mut solver = Solver::default();
 //! # let x = solver.new_bounded_integer(5, 10);
@@ -141,7 +144,13 @@
 //! # let mut termination = Indefinite;
 //! # let mut brancher = solver.default_brancher();
 //! // Then we solve to optimality
-//! let result = solver.minimise(&mut brancher, &mut termination, objective);
+//! let result = solver.optimise(
+//!     &mut brancher,
+//!     &mut termination,
+//!     objective,
+//!     OptimisationDirection::Minimise,
+//!     UpperBoundingSearch,
+//! );
 //!
 //! if let OptimisationResult::Optimal(optimal_solution) = result {
 //!     let value_x = optimal_solution.get_integer_value(x);
@@ -284,9 +293,11 @@ pub(crate) mod basic_types;
 pub mod containers;
 pub(crate) mod engine;
 pub(crate) mod math;
+pub(crate) mod optimisation_search;
 pub(crate) mod propagators;
 pub(crate) mod pumpkin_asserts;
 pub(crate) mod variable_names;
+
 #[cfg(doc)]
 use crate::branching::Brancher;
 #[cfg(doc)]
